+++ PGKStar.swift
@@ -1,6 +1,6 @@
 //
 //  PerseusGeoKitStar.swift
-//  Version: 1.0.0
+//  Version: 1.0.1
 //
 //  Standalone PerseusGeoKit
 //
@@ -212,8 +212,17 @@ public class GeoAgent: NSObject {
 
         log.message("[\(type(of: self))].\(#function) status: \(status)", .notice)
 
+#if os(iOS)
+
+        guard status == .notDetermined else {
+            log.message("[\(type(of: self))].\(#function) status: \(status)", .notice)
+            actionIfAlreadyDetermined?(status)
+            return
+        }
+
+#elseif os(macOS)
+
         guard status == .notDetermined, isAuthorizedForLocationServices == false else {
-#if os(macOS)
             if status == .notDetermined, isAuthorizedForLocationServices {
 
                 // HOTFIX: Location Services Status in OpenCore usage case.
@@ -224,10 +233,12 @@ public class GeoAgent: NSObject {
 
             status = geoStatus
             log.message("[\(type(of: self))].\(#function) status: \(status)", .notice)
-#endif
+
             actionIfAlreadyDetermined?(status)
             return
         }
+
+#endif
 
 #if os(iOS)
 
@@ -250,7 +261,6 @@ public class GeoAgent: NSObject {
         }
 
         locationManager.requestAlwaysAuthorization()
-
 #endif
 
     }
@@ -677,13 +687,13 @@ public struct GeoPoint: CustomStringConvertible, Equatable {
 
     public var description: String {
         /*
-        let locationTwo = "[\(latitude.cut(.two)), \(longitude.cut(.two))]"
+         let locationTwo = "[\(latitude.cut(.two)), \(longitude.cut(.two))]"
 
-        let latitudeFour = "latitude = \(latitude.cut(.four))"
-        let longitudeFour = "longitude = \(longitude.cut(.four))"
+         let latitudeFour = "latitude = \(latitude.cut(.four))"
+         let longitudeFour = "longitude = \(longitude.cut(.four))"
 
-        return locationTwo + ": \(latitudeFour), \(longitudeFour)"
-        */
+         return locationTwo + ": \(latitudeFour), \(longitudeFour)"
+         */
         return "\(latitude.cut(.four)), \(longitude.cut(.four))"
     }