+++ PGKSupportingStar.swift
@@ -1,6 +1,6 @@
 //
 //  PGKSupportingStar.swift
-//  Version: 1.0.0
+//  Version: 1.0.1
 //
 //  PerseusGeoKit Support Code
 //