+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Dates in this file meets Gregorian calendar. Date in format YYYY-MM-DD.
 
+## [1.0.1] - [2025-05-31], PGK
+
+### Fixed
+
+- iOS: Location Services permission request if status in .notDetermined.
+
 ## [1.0.0] - [2025-05-31], PGK
 
 - Minimum build requirements: macOS 10.13+, iOS 11.0+, Xcode 14.2+. If standalone Xcode 10.1+.